+++ src/pthreadpool.c
@@ -227,7 +227,7 @@ struct pthreadpool* pthreadpool_create(size_t threads_count) {
 	return threadpool;
 }
 
-uint32_t pthreadpool_get_threads_count(struct pthreadpool* threadpool) {
+size_t pthreadpool_get_threads_count(struct pthreadpool* threadpool) {
 	return threadpool->threads_count;
 }
